WebServer/requirements.txt
aniso8601==8.0.0
Cerberus==1.3.2
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.18.5
opencv-python==4.5.1.48
pandas==1.0.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
waitress==1.4.4
Werkzeug==1.0.1
